# CI: Fix package benchmarking failing on `next`