--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -239,7 +239,7 @@ jobs:
           steps:
             - run:
                 name: Benchmarking packages against base branch
-                command: yarn bench-packages --baseBranch << pipeline.parameters.ghBaseBranch >> --pull-request << pipeline.parameters.ghPrNumber >> --upload
+                command: yarn bench-packages --base-branch << pipeline.parameters.ghBaseBranch >> --pull-request << pipeline.parameters.ghPrNumber >> --upload
       # if there is a NOT a base branch OR NOT a PR number in parameters, just upload benchmarks for the branch
       # this happens when runned directly on branches, like next or main
       - when: