--- a/.circleci/config.yml
+++ b/.circleci/config.yml
@@ -220,15 +220,6 @@ jobs:
           clone_options: "--depth 1 --verbose"
       - attach_workspace:
           at: .
-      - run:
-          name: Starting local registry
-          working_directory: scripts
-          command: |
-            yarn local-registry --open &
-            until curl -s http://localhost:6001 > /dev/null; do
-              echo 'Waiting for local registry to be available...'
-              sleep 2
-            done
       # if there is a base branch AND a PR number in parameters, benchmark packages against those
       # this happens when run against a PR
       - when:
@@ -240,7 +231,13 @@ jobs:
             - run:
                 name: Benchmarking packages against base branch
                 working_directory: scripts
-                command: yarn bench-packages --base-branch << pipeline.parameters.ghBaseBranch >> --pull-request << pipeline.parameters.ghPrNumber >> --upload
+                command: |
+                  yarn local-registry --open &
+                  until curl -s http://localhost:6001 > /dev/null; do
+                    echo 'Waiting for local registry to be available...'
+                    sleep 2
+                  done
+                  yarn bench-packages --base-branch << pipeline.parameters.ghBaseBranch >> --pull-request << pipeline.parameters.ghPrNumber >> --upload
       # if there is a NOT a base branch OR NOT a PR number in parameters, just upload benchmarks for the branch
       # this happens when runned directly on branches, like next or main
       - when:
@@ -252,7 +249,13 @@ jobs:
             - run:
                 name: Uploading package benchmarks for branch
                 working_directory: scripts
-                command: yarn bench-packages --upload
+                command: |
+                  yarn local-registry --open &
+                  until curl -s http://localhost:6001 > /dev/null; do
+                    echo 'Waiting for local registry to be available...'
+                    sleep 2
+                  done
+                  yarn bench-packages --upload
       - store_artifacts: 
           path: bench/packages/results.json
       - store_artifacts: